day5

+++ src/views/article/index.vue
@@ -5,22 +5,33 @@
       <div slot="header" class="clearfix">
         <span>筛选条件</span>
       </div>
-      <el-form ref="form" :model="form" label-width="80px">
-        <el-form-item label="特殊资源">
-          <el-radio-group v-model="form.resource">
-            <el-radio label="线上品牌商赞助"></el-radio>
-            <el-radio label="线下场地免费"></el-radio>
+      <el-form ref="form" :model="filterParams" label-width="80px">
+        <el-form-item label="状态">
+          <el-radio-group v-model="filterParams.status">
+            <el-radio label="">全部</el-radio>
+            <el-radio
+              v-for="(item, index) in statTypes"
+              :key="item.label"
+              :label="index + ''"
+            >{{ item.label }}</el-radio>
           </el-radio-group>
         </el-form-item>
-        <el-form-item label="活动区域">
-          <el-select v-model="form.region" placeholder="请选择活动区域">
-            <el-option label="区域一" value="shanghai"></el-option>
-            <el-option label="区域二" value="beijing"></el-option>
+        <el-form-item label="频道">
+          <el-select v-model="filterParams.channel_id" placeholder="请选择活动区域">
+            <el-option label="全部" value=""></el-option>
+            <el-option
+              v-for="item in channels"
+              :key="item.id"
+              :label="item.name"
+              :value="item.id"
+            ></el-option>
           </el-select>
         </el-form-item>
-        <el-form-item label="活动形式">
+        <el-form-item label="时间">
           <el-date-picker
-            v-model="form.value1"
+            value-format="yyyy-MM-dd"
+            v-model="begin_end_pubdate"
+            @change="handleDateChange"
             type="daterange"
             range-separator="至"
             start-placeholder="开始日期"
@@ -28,7 +39,11 @@
           </el-date-picker>
         </el-form-item>
         <el-form-item>
-          <el-button type="primary" @click="onSubmit">查询</el-button>
+          <el-button
+            type="primary"
+            :loading="articleLoading"
+            @click="onSubmit"
+          >查询</el-button>
         </el-form-item>
       </el-form>
     </el-card>
@@ -37,7 +52,7 @@
     <!-- 列表 -->
     <el-card class="list-card">
       <div slot="header" class="clearfix">
-        <span>共找到15条符合条件的内容</span>
+        <span>共找到<strong>{{ totalCount }}</strong>条符合条件的内容</span>
       </div>
 
       <!-- table 表格 -->
@@ -78,11 +93,13 @@
           width="180">
         </el-table-column>
         <el-table-column
-          prop="status"
           label="状态">
+          <template slot-scope="scope">
+            <el-tag :type="statTypes[scope.row.status].type">{{ statTypes[scope.row.status].label }}</el-tag>
+          </template>
         </el-table-column>
         <el-table-column
-          label="状态">
+          label="操作">
           <template slot-scope="scope">
             <el-button type="success" plain>修改</el-button>
             <el-button type="danger" plain @click="handleDelete(scope.row)">删除</el-button>
@@ -96,6 +113,8 @@
         一：分多少页
           每页多大，默认是10条每页，我们的接口如果没有指定每页条数，则默认也是按照每页10条返回数据
           有多少条数据
+          total  总记录数
+          current-page 当前页码，也就是高亮的那个页码
         二：页面改变加载对应的页码数据
        -->
       <el-pagination
@@ -103,6 +122,7 @@
         layout="prev, pager, next"
         :total="totalCount"
         :disabled="articleLoading"
+        :current-page="page"
         @current-change="handleCurrentChange"
       >
       </el-pagination>
@@ -118,36 +138,71 @@ export default {
   data () {
     return {
       articles: [], // 列表数据
-      form: {
-        name: '',
-        region: '',
-        date1: '',
-        date2: '',
-        delivery: false,
-        type: [],
-        resource: '',
-        desc: '',
-        value1: ''
+      totalCount: 0, // 总数据条数
+      articleLoading: false, // 控制文档加载中的 loading 效果
+      page: 1, // 当前页码
+      statTypes: [ // 文章状态
+        {
+          type: 'info',
+          label: '草稿'
+        },
+        {
+          type: '',
+          label: '待审核'
+        },
+        {
+          type: 'success',
+          label: '审核通过'
+        },
+        {
+          type: 'warning',
+          label: '审核失败'
+        },
+        {
+          type: 'danger',
+          label: '已删除'
+        }
+      ],
+      channels: [], // 频道列表
+      filterParams: { // 文章查询条件参数
+        status: '', // 文章状态
+        channel_id: '', // 频道id
+        begin_pubdate: '', // 开始时间
+        end_pubdate: '' // 结束时间
       },
-      totalCount: 0,
-      articleLoading: false,
-      page: 1
+      begin_end_pubdate: [] // 存储日期选择器同步的 [开始时间，结束时间]，这个字段没啥用，只是日期选择器必须 v-mode 绑定一个数据才会触发 change 事件
     }
   },
 
   created () {
+    // 加载文章列表
     this.loadArticles()
+
+    // 加载频道列表
+    this.loadChannels()
   },
 
   methods: {
     loadArticles (page = 1) { // 函数参数的默认值
       this.articleLoading = true
+
+      // 过滤出有效的查询条件数据字段
+      const filterData = {}
+      for (let key in this.filterParams) {
+        if (this.filterParams[key]) {
+          filterData[key] = this.filterParams[key]
+        }
+      }
+
+      console.log(filterData)
+
       this.$http({
         method: 'GET',
         url: '/articles',
         params: {
           page, // 请求数据的页码，不传默认为 1
-          per_page: 10// 请求数据的每页大小，不传默认为 10
+          per_page: 10, // 请求数据的每页大小，不传默认为 10
+          ...filterData // 将对象混入当前对象，说白就是对象拷贝
         }
       }).then(data => {
         this.articles = data.results // 列表数据
@@ -156,8 +211,18 @@ export default {
       })
     },
 
+    loadChannels () {
+      this.$http({
+        method: 'GET',
+        url: '/channels'
+      }).then(data => {
+        this.channels = data.channels
+      })
+    },
+
     onSubmit () {
-      console.log('submit!')
+      this.page = 1 // 让分页组件的页码回到第1页
+      this.loadArticles() // 加载第1页的数据
     },
 
     handleCurrentChange (page) {
@@ -167,12 +232,39 @@ export default {
     },
 
     handleDelete (article) {
-      this.$http({
-        method: 'DELETE',
-        url: `/articles/${article.id}`
-      }).then(data => {
-        this.loadArticles(this.page)
+      this.$confirm('确认删除吗？', '删除提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => { // 确定执行
+        // 发送删除请求
+        this.$http({
+          method: 'DELETE',
+          url: `/articles/${article.id}`
+        }).then(data => {
+          // 提示删除成功
+          this.$message({
+            type: 'success',
+            message: '删除成功!'
+          })
+
+          // 重新加载数据列表
+          this.loadArticles(this.page)
+        })
+      }).catch(() => { // 取消执行
+        this.$message({
+          type: 'info',
+          message: '已取消删除'
+        })
       })
+    },
+
+    /**
+     * 日期选择组件改变事件
+     */
+    handleDateChange (value) {
+      this.filterParams.begin_pubdate = value[0]
+      this.filterParams.end_pubdate = value[1]
     }
   }
 }
